+++ frontpage-engine.php
@@ -5,7 +5,7 @@
  * Description: A Maverick Engine plugin that brings order to your Wordpress front page.
  * Author: Daily Maverick, Jason Norwood-Young, Realm Digital
  * Author URI: https://dailymaverick.co.za
- * Version: 0.1.1
+ * Version: 0.2.5
  * License: GPLv3
  * License URI: https://www.gnu.org/licenses/gpl-3.0.html
  * WC requires at least: 5.8.0
@@ -17,7 +17,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-const FRONTPAGEENGINE_VERSION = '0.1.1';
+const FRONTPAGEENGINE_VERSION = '0.2.5';
 $frontpageengine_menu_slug = "frontpage-engine-menu";
 
 require_once(plugin_dir_path( __FILE__ ).'includes/admin/tables/frontpage_engine-frontpages_table.php');
@@ -65,14 +65,9 @@ function frontpage_engine_database_setup() {
 add_action( 'init', 'frontpage_engine_database_setup', 2 );
 
 function frontpage_engine_serve_init() {
-    require_once(plugin_dir_path( __FILE__ ) . 'includes/frontpage_engine-serve.php' );
-    // Serve frontpage if url matches /frontpage-engine/serve/{frontpage_name}
+    require_once(plugin_dir_path( __FILE__ ) . 'includes/serve/frontpage_engine-serve.php' );
+    // Serve frontpage if url matches /frontpage-engine/serve/{frontpage_slug}
     if (empty($_SERVER['REQUEST_URI'])) return;
-    if (preg_match('/\/frontpage-engine\/serve\/(.*)/', $_SERVER['REQUEST_URI'], $matches)) {
-        $frontpage_name = $matches[1];
-        $frontpage_engine_serve = new FrontpageEngineServe();
-        $frontpage_engine_serve->serve($frontpage_name, 'json');
-        exit;
-    }
+    new FrontpageEngineServe();
 }
 add_action( 'init', 'frontpage_engine_serve_init' );